# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  backend_internal_domain_model.Document:
    properties:
      author:
        $ref: '#/definitions/backend_internal_domain_model.UserAuthor'
      authorId:
        type: string
      categories:
        items:
          type: string
        type: array
      category:
        description: Primary / first category
        type: string
      content:
        type: string
      createdAt:
        type: string
      deletedAt:
        type: string
      deletedBy:
        type: string
      deletedByUser:
        $ref: '#/definitions/backend_internal_domain_model.UserAuthor'
      id:
        type: string
      isDraft:
        type: boolean
      isShared:
        type: boolean
      isStarred:
        type: boolean
      lastViewedAt:
        type: string
      projectId:
        type: string
      projectName:
        type: string
      starredAt:
        type: string
      tags:
        items:
          type: string
        type: array
      thumbnail:
        type: string
      thumbnailDark:
        type: string
      thumbnailPreview:
        type: string
      thumbnailPreviewDark:
        type: string
      title:
        type: string
      type:
        description: markdown, dbdiagram, mermaid
        type: string
      updatedAt:
        type: string
      viewCount:
        type: integer
      visibility:
        description: workspace, private, public_link, inherit
        type: string
      workspaceId:
        type: string
    type: object
...
  backend_internal_domain_model.UserAuthor:
    properties:
      avatar:
        type: string
      email:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
...
  backend_internal_presentation_document_presenter.CreateDocumentRequest:
    properties:
      categories:
        items:
          type: string
        type: array
      content:
        type: string
      initialBody:
        $ref: '#/definitions/backend_internal_presentation_document_presenter.CreateMarkdownBodyRequest'
      isDraft:
        type: boolean
      projectId:
        type: string
      tags:
        items:
          type: string
        type: array
      title:
        type: string
      type:
        type: string
      visibility:
        type: string
    type: object
  backend_internal_presentation_document_presenter.CreateMarkdownBodyNode:
    properties:
      attributes:
        items:
          type: integer
        type: array
      content:
        type: string
      nodeID:
        type: string
      parentID:
        type: string
      siblingOrder:
        type: number
      type:
        type: string
    type: object
  backend_internal_presentation_document_presenter.CreateMarkdownBodyRequest:
    properties:
      bodySchemaVersion:
        type: integer
      documentID:
        type: string
      nodes:
        items:
          $ref: '#/definitions/backend_internal_presentation_document_presenter.CreateMarkdownBodyNode'
        type: array
      rootNodeID:
        type: string
    type: object
  backend_internal_presentation_document_presenter.MoveRequest:
    properties:
      targetProjectId:
        type: string
    type: object
...
  backend_internal_presentation_response.Envelope:
    properties:
      data: {}
    type: object
  backend_internal_presentation_response.ErrorEnvelope:
    properties:
      title:
        type: string
    type: object
...
  internal_presentation_document_handler.bodyNodeResponse:
    properties:
      attributes:
        items:
          type: integer
        type: array
      content:
        type: string
      nodeID:
        type: string
      parentID:
        type: string
      siblingOrder:
        type: number
      type:
        type: string
      version:
        type: integer
    type: object
...
  internal_presentation_document_handler.publicBodySnapshotResponse:
    properties:
      bodySchemaVersion:
        type: integer
      bodyVersion:
        type: integer
      nodes:
        items:
          $ref: '#/definitions/internal_presentation_document_handler.bodyNodeResponse'
        type: array
      rootNodeID:
        type: string
    type: object
info:
  contact: {}
  description: API documentation for Dokudocs backend service.
  title: Dokudocs API
  version: "1.0"
paths:
...
  /documents:
    get:
      description: Get list of documents in the active workspace
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      - description: Filter tab (e.g. all, recent, starred, draft)
        in: query
        name: filterTab
        type: string
      - description: Search term
        in: query
        name: search
        type: string
      - description: Category filter
        in: query
        name: category
        type: string
      - description: Sort field (e.g. updated_at, title)
        in: query
        name: sortField
        type: string
      - description: Sort order (asc, desc)
        in: query
        name: sortOrder
        type: string
      - description: Filter by project ID (UUID)
        in: query
        name: projectId
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/backend_internal_domain_model.Document'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: List documents
      tags:
      - Document
    post:
      consumes:
      - application/json
      description: Create a document. Markdown may include an initial AST body stored
        atomically with its metadata.
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      - description: Stable request ID (UUID); reuse on retry
        in: header
        name: Idempotency-Key
        required: true
        type: string
      - description: Document creation payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_internal_presentation_document_presenter.CreateDocumentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/backend_internal_domain_model.Document'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Create document
      tags:
      - Document
...
  /documents/{id}/accesses/{accessId}:
    delete:
      description: Remove collaborator access from a document
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      - description: Document ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Collaborator User ID (UUID)
        in: path
        name: accessId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Revoke document access
      tags:
      - Document
...
  /documents/{id}/move:
    put:
      consumes:
      - application/json
      description: Move document to a target project or null for root
      parameters:
      - description: Workspace ID (UUID)
        in: header
        name: X-Workspace-Id
        required: true
        type: string
      - description: Document ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Move destination payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/backend_internal_presentation_document_presenter.MoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  additionalProperties:
                    type: string
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      security:
      - BearerAuth: []
      summary: Move document
      tags:
      - Document
...
  /public/documents/{shareToken}/body:
    get:
      description: Returns a shared document's canonical AST without exposing its
        CRDT state.
      parameters:
      - description: Document Share Token
        in: path
        name: shareToken
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/backend_internal_presentation_response.Envelope'
            - properties:
                data:
                  $ref: '#/definitions/internal_presentation_document_handler.publicBodySnapshotResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/backend_internal_presentation_response.ErrorEnvelope'
      summary: Get public Markdown body
      tags:
      - Document
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"